move styling and components to new components dir
