+++ src/main.jsx
@@ -4,7 +4,7 @@ import {
   createBrowserRouter,
   RouterProvider,
 } from 'react-router-dom';
-import Root from './routes/root';
+import Root from './root';
 import ErrorPage from './error';
 
 const router = createBrowserRouter([
