+++ src/root.jsx
@@ -1,6 +1,6 @@
-import Header from './header';
-import Body from './body';
-import Footer from './footer'
+import Header from './components/header';
+import Body from './components/body';
+import Footer from './components/footer'
 
 function Root() {
     return (
